Always complete when explicitly selected
Previously, TAB would only select a completion if there was already
non-whitespace text on the line.  This was so that Copilot suggestions
did not interfere with using TAB to change indentation.  However, it is
also useful to allow Copilot suggestions to be explicitly selected using
arrow keys, and, in that case, we should complete on TAB even if there
is no other text on the line.